requirements.txt
Django==5.0
mysqlclient==2.2.1
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
spacy==3.7.2
nltk==3.8.1
python-magic==0.4.27
gunicorn==21.2.0
python-dotenv==1.0.0
# For production deployment
whitenoise==6.6.0 